+++ tests/TODO
@@ -1,0 +1,34 @@
+tools.py
+--------
+- maybe make stub Execer a fixture
+  for (text_execer, test_ast, test_context)
+
+test_tools.py
+-------------
+- EnvPath bug
+- refactor test_partial_string
+- refactor test_executable
+- maybe split test_commands_cache_lazy
+
+
+test_aliases.py
+---------------
+- migrate TestWhich
+
+
+test_builtins.py
+test_contexts.py
+test_dirstack.py
+test_environ.py
+test_execer.py
+test_foreign_shells.py
+test_history.py
+test_imphooks.py
+test_lazyjson.py
+test_lexer.py
+test_main.py
+test_parser.py
+test_ptk_multiline.py
+test_replay.py
+test_tools.py
+test_wizard.py
